+++ e2e/node/acp.test.ts
@@ -62,14 +62,14 @@ describe("An ACP Solid server", () => {
   let options: { fetch: typeof global.fetch };
   let session: Session;
   let sessionResource: string;
-  
+
   beforeEach(async () => {
     session = await getAuthenticatedSession(env);
     sessionResource = `${env.pod}${sessionResourcePrefix}${session.info.sessionId}`;
     options = { fetch: session.fetch };
     await saveSolidDatasetAt(sessionResource, createSolidDataset(), options);
   });
-  
+
   afterEach(async () => {
     await deleteSolidDataset(sessionResource, options);
     await session.logout();
@@ -117,13 +117,15 @@ describe("An ACP Solid server", () => {
       controlRead: false,
       controlWrite: false,
     });
-    expect(await getAgentAccess(sessionResource, agent, options)).toStrictEqual({
-      read: true,
-      append: false,
-      write: false,
-      controlRead: false,
-      controlWrite: false,
-    });
+    expect(await getAgentAccess(sessionResource, agent, options)).toStrictEqual(
+      {
+        read: true,
+        append: false,
+        write: false,
+        controlRead: false,
+        controlWrite: false,
+      }
+    );
   });
 
   it("can get and set read access for the public", async () => {
@@ -136,9 +138,9 @@ describe("An ACP Solid server", () => {
         ? latest_getPublicAccess
         : legacy_getPublicAccess;
 
-    await expect(
-      getSolidDataset(sessionResource, options)
-    ).resolves.toEqual(expect.objectContaining({ graphs: { default: {} } }));
+    await expect(getSolidDataset(sessionResource, options)).resolves.toEqual(
+      expect.objectContaining({ graphs: { default: {} } })
+    );
 
     await expect(getSolidDataset(sessionResource)).rejects.toThrow();
 
@@ -196,13 +198,15 @@ describe("An ACP Solid server", () => {
       controlRead: true,
       controlWrite: true,
     });
-    expect(await getAgentAccess(sessionResource, agent, options)).toStrictEqual({
-      read: true,
-      append: true,
-      write: true,
-      controlRead: true,
-      controlWrite: true,
-    });
+    expect(await getAgentAccess(sessionResource, agent, options)).toStrictEqual(
+      {
+        read: true,
+        append: true,
+        write: true,
+        controlRead: true,
+        controlWrite: true,
+      }
+    );
   });
 
   it("can remove access for an agent", async () => {
@@ -238,12 +242,14 @@ describe("An ACP Solid server", () => {
       controlRead: false,
       controlWrite: false,
     });
-    expect(await getAgentAccess(sessionResource, agent, options)).toStrictEqual({
-      read: true,
-      append: true,
-      write: false,
-      controlRead: false,
-      controlWrite: false,
-    });
+    expect(await getAgentAccess(sessionResource, agent, options)).toStrictEqual(
+      {
+        read: true,
+        append: true,
+        write: false,
+        controlRead: false,
+        controlWrite: false,
+      }
+    );
   });
 });